refactor(reborn): isolate production runtime wiring
